refactor: ResultRestaurantCard 컴포넌트 분리 및 디렉터리 도메인 폴더로 이동
